# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kakusin API",
        "description": "Tài liệu API Kakusin",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Danh sách News (Tổng)",
                "operationId": "9bfe744cf19dcb58e9ab9677c29ee6c7",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "ja"
                        }
                    },
                    {
                        "name": "featured",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công"
                    }
                }
            }
        },
...
        "/api/news/{slug}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Chi tiết News",
                "operationId": "6f16ebaf3fe9643317a035fba8c160ea",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "ja"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy bài viết"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Danh sách Product (Tổng/Phân trang)",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "ja"
                        }
                    },
                    {
                        "name": "featured",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "all"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công"
                    }
                }
            }
        },
        "/api/products/{slug}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Chi tiết Product",
                "operationId": "2eb94cf22f00aa181040441a8ae7b07e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "ja"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu chi tiết thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy sản phẩm"
                    }
                }
            }
        },
...
    },
...
}